Changes in the implementation of custom classes
A new wrapper has been defined (GeneralPorduct) while the others have been modified. The idea behind it was to allow for cached accesses to the matrix element. Yet the computation of the elements of the local matrix seem to be more efficient if computed using some temporary Eigen matrixes. Some research still needs to be done